// src/lib/schemas/tables/multi-simulation-table-schema.ts
export const multiSimulationTableRowSchema = z.object({
  seed: z.number(),
  success: z.boolean(),
  retirementAge: z.number().nullable(),
  bankruptcyAge: z.number().nullable(),
  finalPhaseName: z.string().nullable(),
  finalPortfolioValue: z.number(),
  minStockReturn: z.number().nullable(),
  maxStockReturn: z.number().nullable(),
  meanStockReturn: z.number().nullable(),
  earlyRetirementStockReturn: z.number().nullable(),
  meanBondReturn: z.number().nullable(),
  meanCashReturn: z.number().nullable(),
  meanInflationRate: z.number().nullable(),
  lifetimeIncomeTax: z.number().nullable(),
  lifetimeFicaTax: z.number().nullable(),
  lifetimeCapGainsTax: z.number().nullable(),
  lifetimeNiit: z.number().nullable(),
  lifetimeEarlyWithdrawalPenalties: z.number().nullable(),
  lifetimeTaxesAndPenalties: z.number().nullable(),
  lifetimeReturns: z.number().nullable(),
  lifetimeContributions: z.number().nullable(),
  lifetimeWithdrawals: z.number().nullable(),
  lifetimeRealizedGains: z.number().nullable(),
  lifetimeRequiredMinimumDistributions: z.number().nullable(),
  historicalRanges: z.array(z.object({ startYear: z.number(), endYear: z.number() })).nullable(),
});

export type MultiSimulationTableRow = z.infer<typeof multiSimulationTableRowSchema>;
// ...
export const yearlyAggregateTableRowSchema = z.object({
  year: z.number(),
  age: z.number(),

  percentAccumulation: z.number(),
  percentRetirement: z.number(),
  percentBankrupt: z.number(),

  p10PortfolioValue: z.number(),
  p25PortfolioValue: z.number(),
  p50PortfolioValue: z.number(),
  p75PortfolioValue: z.number(),
  p90PortfolioValue: z.number(),
});

export type YearlyAggregateTableRow = z.infer<typeof yearlyAggregateTableRowSchema>;